fga-eps-mds/2018.2-IndicaAi#256 - Merge with indica-ai devel and change location searchscreen
