--- mobile/screens/tab_navigator/indica_ai/RootStack.js
@@ -1,16 +1,0 @@
-import {StackNavigator} from 'react-navigation';
-
-import SearchScreen from './screens/SearchScreen.js'
-import ViewLocal from './containers/ViewLocal'
-
-const RootStack = StackNavigator({
-
-    Home :{
-      screen: SearchScreen,
-    },
-    ViewLocal : {
-      screen: ViewLocal,
-    }
-})
-
-export default RootStack;
