+++ mobile/screens/tab_navigator/indica_ai/containers/ViewLocal.js
@@ -15,17 +15,14 @@ width = Dimensions.get('window').width;
 export default class ViewLocal extends Component{
     constructor(props) {
       super(props);
-      // this.state = {
-      //   local: props.navigation.state.params.local,
-      // }
     }
   render() {
-    console.log("Teste Inicio !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!")
-    console.log(this.props);
-    console.log("Teste Fim !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!")
     const {state} = this.props.navigation;
     var local = state.params ? state.params.local : undefined;
+    var id = local.id;
     var name = local.name;
+    var description = local.description;
+    
     return (
         <View style={styles.container}>
           <ScrollView>
@@ -34,8 +31,6 @@ export default class ViewLocal extends Component{
             />
             <View style={styles.localContainer}>
               <Text style={styles.localName}>
-                {/* Local Name */}
-                {/* {this.state.local.name} */}
                 {name}
               </Text>
               <Icon style={styles.localHeart}
