+++ mobile/screens/tab_navigator/indica_ai/screens/SearchScreen.js
@@ -1,36 +1,16 @@
-import React, { Component } from "react";
-import {
-  View,
-  StyleSheet
-} from "react-native";
-import { Spinner } from 'native-base';
-import { connect } from 'react-redux';
-import SearchBar from "../containers/searchBar";
-import ListLocals from '../containers/ListLocals'
+import {StackNavigator} from 'react-navigation';
 
+import ViewLocal from '../containers/ViewLocal';
+import SearchLocal from './search_local/SearchLocal';
 
-class SearchScreen extends Component {
-  
-  render() {
-      return (
-          <View style={styles.container}>
-              <SearchBar />
-              <ListLocals />
-          </View>
-      );
-  }
-}
+const SearchScreen = StackNavigator({
+
+    Home :{
+      screen: SearchLocal,
+    },
+    ViewLocal : {
+      screen: ViewLocal,
+    }
+})
 
 export default SearchScreen;
-
-const styles = StyleSheet.create({
-
-  container: {
-    flex: 1,
-    flexDirection: "column",
-    justifyContent: "flex-start",
-    backgroundColor: "#FFF",
-    paddingHorizontal: 10,
-  }
-
-});
