+++ mobile/screens/tab_navigator/indica_ai/screens/search_local/SearchLocal.js
@@ -1,0 +1,36 @@
+import React, { Component } from "react";
+import {
+  View,
+  StyleSheet
+} from "react-native";
+import { Spinner } from 'native-base';
+import { connect } from 'react-redux';
+import SearchBar from "../../containers/searchBar";
+import ListLocals from '../../containers/ListLocals'
+
+
+class SearchLocal extends Component {
+  
+  render() {
+      return (
+          <View style={styles.container}>
+              <SearchBar />
+              <ListLocals />
+          </View>
+      );
+  }
+}
+
+export default SearchLocal;
+
+const styles = StyleSheet.create({
+
+  container: {
+    flex: 1,
+    flexDirection: "column",
+    justifyContent: "flex-start",
+    backgroundColor: "#FFF",
+    paddingHorizontal: 10,
+  }
+
+});
